% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{gen_example_cohort}
\alias{gen_example_cohort}
\title{Title}
\usage{
gen_example_cohort(filename, seed = 1, n = 50)
}
\arguments{
\item{filename}{d}

\item{seed}{d}

\item{n}{d}
}
\value{
d
}
\description{
Title
}
