% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrePost.R
\name{generate_theographs,PrePost-method}
\alias{generate_theographs,PrePost-method}
\title{generate_theographs}
\usage{
\S4method{generate_theographs}{PrePost}(x)
}
\arguments{
\item{PrePost}{PrePost object}
}
\value{
a figure
}
\description{
generate_theographs
}
