% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrePost.R
\name{generate_theographs}
\alias{generate_theographs}
\title{generate_theographs}
\usage{
generate_theographs(obj)
}
\arguments{
\item{obj}{PrePost object}
}
\value{
a figure
}
\description{
generate_theographs
}
