% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrePost.R
\name{generate_trace_plots}
\alias{generate_trace_plots}
\title{generate_trace_plots}
\usage{
generate_trace_plots(x, period)
}
\arguments{
\item{x}{PrePost object}

\item{period}{one of c("after", "before", "around")}
}
\value{
a figure
}
\description{
generate_trace_plots
}
