% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrePost.R
\name{get_window,PrePost-method}
\alias{get_window,PrePost-method}
\title{get_window}
\usage{
\S4method{get_window}{PrePost}(x)
}
\arguments{
\item{x}{object}
}
\value{
named list with the window
}
\description{
get_window
}
