% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrePost.R
\docType{class}
\name{PrePost-class}
\alias{PrePost-class}
\alias{PrePost}
\title{PrePost}
\value{
PrePost object
}
\description{
PrePost
}
\section{Slots}{

\describe{
\item{\code{nhs_number}}{character.}

\item{\code{index_event_time}}{POSIXct.}

\item{\code{window_pre}}{d}

\item{\code{window_post}}{d}

\item{\code{window_units}}{d}

\item{\code{activity_filter}}{d}

\item{\code{svr_name}}{d}

\item{\code{svr}}{d}
}}

