% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{prepost_sens}
\alias{prepost_sens}
\title{Run sensitivity analysis for the randomized moderator placement design}
\usage{
prepost_sens(
  formula,
  data,
  moderator,
  prepost,
  g_by,
  t_by,
  sims = 1000,
  conf_level = 0.95,
  moderator_mono = NULL,
  outcome_mono = NULL,
  c_se = NULL,
  progress = TRUE
)
}
\arguments{
\item{formula}{A formula with syntax \code{y ~ t}, where \code{y} is the
(unquoted) name of the outcome and \code{t} is the (unquoted) name of the treatment.}

\item{data}{A data.frame containing variables in the formula, moderator, and covariates arguments.}

\item{moderator}{A one-sided formuala with syntax \code{~ d}, where \code{d}
is the (unquoted) name of the  moderator variable for the CATE.}

\item{prepost}{A one-sided formula with syntax ~ z, where z is the indicator variable for whether the moderator was measured pre- or post-treatment.}

\item{g_by}{Numeric indicating the grid spacing for the
\eqn{\gamma} parameter that places an upper bound on the proportion
of units whose moderator is affected by treatment.}

\item{t_by}{Numeric indicating the grid spacing for the
\eqn{\theta} parameter that restricts what proportion of units have
their outcomes affected by the pre vs post-measurement of the
moderator.}

\item{sims}{An integer indicating the number of simulations for the
bootstrap confidence intervals for the bounds.}

\item{conf_level}{A numeric indicating the confidence level for the
bootstrap confidence intervals.}

\item{moderator_mono}{A integer or vector of length 2 indicating
if the bounds should assume monotonicity of the effect of the
post-test on the moderator with \code{1} indicating that the post-test
effect is positive and \code{-1} indicating that it is negative. The
vector of length 2 allows the monotonicity assumption to vary by
treatment status with the first entry being for control and the
second for treated.}

\item{outcome_mono}{A integer or vector of length 2 indicating
if the bounds should assume monotonicity of the effect of the
post-test on the outcome with \code{1} indicating that the post-test
effect is positive and \code{-1} indicating that it is negative. The
vector of length 2 allows the monotonicity assumption to vary by
treatment status with the first entry being for control and the
second for treated.}

\item{c_se}{A numeric vector of positive values that allow for the
bounds to be estimated under a relaxation of \code{c_se[i]} standard
errors of the population constraints. Bounds are returned for the
smallest value of \code{c_se} where bounds are feasible. By default,
this is set to a grid of \code{seq(0, 3, 0.1)}.}

\item{progress}{A logical indicating if progress bars should be
displayed. Defaults to TRUE.}
}
\value{
A list object containing sensitivity output.
}
\description{
Run sensitivity analysis for the randomized moderator placement design
}
\examples{
x <- "alfa,bravo,charlie,delta"
}
