% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrePost.R
\name{run_activity_summary,PrePost-method}
\alias{run_activity_summary,PrePost-method}
\title{run_activity_summary}
\usage{
\S4method{run_activity_summary}{PrePost}(x, period = "bf")
}
\arguments{
\item{period}{either bf or af (before or after)}

\item{PrePost}{PrePost object}
}
\value{
a table
}
\description{
run_activity_summary
}
