% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrePost.R
\name{run_activity_summary}
\alias{run_activity_summary}
\title{run_activity_summary}
\usage{
run_activity_summary(x, period)
}
\arguments{
\item{x}{PrePost object}

\item{period}{one of c("after", "before")}
}
\value{
a table
}
\description{
run_activity_summary
}
