% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrePost.R
\name{run_activity_volume}
\alias{run_activity_volume}
\title{run_activity_volume}
\usage{
run_activity_volume(x, period)
}
\arguments{
\item{x}{PrePost object}

\item{period}{one of c("after", "before", "around")}
}
\value{
a table
}
\description{
run_activity_volume
}
