% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrePost.R
\name{run_descriptives,PrePost-method}
\alias{run_descriptives,PrePost-method}
\title{Title}
\usage{
\S4method{run_descriptives}{PrePost}(x)
}
\arguments{
\item{PrePost}{PrePost object}
}
\value{
a figure
}
\description{
Title
}
