% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrePost.R
\name{set_window,PrePost-method}
\alias{set_window,PrePost-method}
\title{set_window}
\usage{
\S4method{set_window}{PrePost}(x, pre, post, units = "days")
}
\arguments{
\item{x}{d}

\item{pre}{d}

\item{post}{d}

\item{units}{d}
}
\value{
the updated object
}
\description{
set_window
}
