% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrePost.R
\name{set_window}
\alias{set_window}
\title{set_window}
\usage{
set_window(x, pre, post, units)
}
\arguments{
\item{x}{d}

\item{pre}{d}

\item{post}{d}

\item{units}{d}
}
\value{
the updated object
}
\description{
set_window
}
