% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrePost.R
\name{generate_trace_plots,PrePost-method}
\alias{generate_trace_plots,PrePost-method}
\title{generate_trace_plots}
\usage{
\S4method{generate_trace_plots}{PrePost}(x)
}
\arguments{
\item{PrePost}{PrePost object}
}
\value{
a figure
}
\description{
generate_trace_plots
}
