% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{prepost_gibbs_nocovar}
\alias{prepost_gibbs_nocovar}
\title{Run Gibbs sampler without covariates}
\usage{
prepost_gibbs_nocovar(
  formula,
  data,
  prepost,
  moderator,
  iter = 1000,
  thin = 1,
  burn = 0,
  offset = 0,
  monotonicity = TRUE,
  stable = TRUE,
  priors,
  predictive = FALSE
)
}
\arguments{
\item{formula}{A formula with syntax y ~ t, where y is the outcome and t is the treatment.}

\item{data}{A dataframe}

\item{prepost}{A formula with syntax ~ z, where z is the indicator variable for whether the moderator was measured pre- or post-treatment...}

\item{moderator}{A formuala with syntax ~ d, where d is the moderator variable for the CATE.}

\item{iter}{Numeric, number of iterations for the Gibbs}

\item{thin}{Numeric, thinning parameter for the Gibbs}

\item{burn}{Numeric, burn in rate for the Gibbs}

\item{offset}{Numeric, ...}

\item{stable}{A logical signifying whether Gibbs assumes stability.}

\item{priors}{A list object containing the priors for the Gibbs sampler. Priors include beta.precision, psi.precision, alpha, y.alpha, and y.beta.}

\item{monotonicty}{A logical signifying whether Gibbs assumes monotonicity.}

\item{saturated}{A logical signifying...}
}
\value{
A list object containing Gibbs posterior quantities of interest and parameters.
}
\description{
Run Gibbs sampler without covariates
}
\examples{
x <- "alfa,bravo,charlie,delta"
}
