% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrePost.R
\name{run_activity_summary,PrePost-method}
\alias{run_activity_summary,PrePost-method}
\title{run_activity_summary}
\usage{
\S4method{run_activity_summary}{PrePost}(x)
}
\arguments{
\item{PrePost}{PrePost object}
}
\value{
a figure
}
\description{
run_activity_summary
}
